feat: Add confirmation prompt to git commit
This commit introduces a confirmation prompt before creating a git commit. If the 'confirm' flag is set, the generated commit header and message will be displayed to the user, who will then be prompted to confirm whether to proceed with the commit. The user can type 'yes' or simply press Enter to confirm, or type 'no' to abort the commit.